fix: fixed typo polution->pollution, use of get_data function for vehicules
